Add files via upload
Invoke-Reprofile renames the UserName to UserName.old in Users Folder, Backup User SID in Registry to Windows Temp Folder, and Renamed SID to SID.Old in ProfileList. If user profile is messed up on multiple systems, use this on multiple computers.

WinRM must be Configured on Remote Computers and Remote Users must be Logged off. I have other functions for doing that. Not included with this example at this time.